+++ api/CHANGELOG.md
@@ -1,12 +1,14 @@
 ## master
 
-_06/19/2023_
+_06/20/2023_
 
 https://github.com/gear-tech/gear-js/pull/1298
+https://github.com/gear-tech/gear-js/pull/1301
 
 ### Changes
 
 - New approach to generate program IDs
+- Support `resumeSessionInit`, `resumeSessionPush` and `resumeSessionCommit` extrinsics according to https://github.com/gear-tech/gear/pull/2622
 
 ## 0.31.2